interface/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "interface-agent-bearing"
version = "0.1.0"
dependencies = [
  "fastapi>=0.95",
  "uvicorn[standard]>=0.22",
  "python-dotenv",
  "librosa",
  "soundfile",
  "numpy",
  "requests",
  "tensorflow",  # or tensorflow-cpu if you prefer
  "scipy"
]
